refactor: change folder name, remove react fragment and comment
